Moved protocol to its own crate in workspace. WIP
