+++ Cargo.toml
@@ -1,9 +1,12 @@
+[workspace]
+members = ["protocol"]
+
 [package]
 name = "slimevr-wrangler"
 version = "0.5.0"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+publish = false
+license = "GPL-3.0"
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
@@ -14,13 +17,17 @@ async-std = "1.0"
 joycon-rs = { git = "https://github.com/carl-anders/joycon-rs", optional = true }
 joycon = { git = "https://github.com/Yamakaky/joy", optional = true }
 directories = "4.0"
-self_update = { version = "0.30", features = ["archive-zip", "compression-zip-deflate"] }
+self_update = { version = "0.30", features = [
+    "archive-zip",
+    "compression-zip-deflate",
+] }
 itertools = "0.10"
 ahrs = "0.5"
 nalgebra = "0.30"
 deku = "0.13"
 md-5 = "0.10"
 arc-swap = "1.5"
+protocol = { path = "protocol" }
 
 [features]
 default = ["use-joycon-rs"]
