--- src/slime/mod.rs
@@ -1,2 +1,0 @@
-pub mod deku;
-mod test_deku;
